+++ src/main/java/uk/gov/pay/ledger/exception/BadRequestExceptionMapper.java
@@ -2,7 +2,7 @@ package uk.gov.pay.ledger.exception;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import uk.gov.pay.commons.model.ErrorIdentifier;
+import uk.gov.service.payments.commons.model.ErrorIdentifier;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
