+++ src/main/java/uk/gov/pay/ledger/exception/ErrorResponse.java
@@ -2,7 +2,7 @@ package uk.gov.pay.ledger.exception;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import uk.gov.pay.commons.model.ErrorIdentifier;
+import uk.gov.service.payments.commons.model.ErrorIdentifier;
 
 import java.util.List;
