+++ src/main/java/uk/gov/pay/ledger/exception/JerseyViolationExceptionMapper.java
@@ -3,7 +3,7 @@ package uk.gov.pay.ledger.exception;
 import io.dropwizard.jersey.validation.JerseyViolationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import uk.gov.pay.commons.model.ErrorIdentifier;
+import uk.gov.service.payments.commons.model.ErrorIdentifier;
 
 import javax.validation.ConstraintViolation;
 import javax.ws.rs.core.Response;
