+++ src/main/java/uk/gov/pay/ledger/transaction/search/model/PaymentSettlementSummary.java
@@ -7,7 +7,7 @@ import java.time.format.DateTimeFormatter;
 import java.util.Objects;
 import java.util.Optional;
 
-import static uk.gov.pay.commons.model.ApiResponseDateTimeFormatter.ISO_INSTANT_MILLISECOND_PRECISION;
+import static uk.gov.service.payments.commons.model.ApiResponseDateTimeFormatter.ISO_INSTANT_MILLISECOND_PRECISION;
 
 public class PaymentSettlementSummary extends SettlementSummary {
     private ZonedDateTime settlementSubmittedTime;
